+++ crates/metadata/src/metadata_error.rs
@@ -14,4 +14,15 @@ pub enum MetadataError {
     #[diagnostic(code(MetadataError::Deserialize), help(""))]
     #[error("toml load failed")]
     Deserialize(#[from] toml::de::Error),
+
+    #[diagnostic(
+        code(MetadataError::InvalidProjectName),
+        help("\"[a-zA-Z_][0-9a-zA-Z_]*\" can be used as project name")
+    )]
+    #[error("project name \"{0}\" is invalid")]
+    InvalidProjectName(String),
+
+    #[diagnostic(code(MetadataError::InvalidLicense), help(""))]
+    #[error("license parse failed")]
+    InvalidLicense(#[from] spdx::ParseError),
 }