+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased](https://github.com/dalance/veryl/compare/v0.2.1...Unreleased) - ReleaseDate
 
 * [Added] attribute arguments check [#90](https://github.com/dalance/veryl/issues/90)
+* [Added] metadata check [#91](https://github.com/dalance/veryl/issues/91)
 
 ## [v0.2.1](https://github.com/dalance/veryl/compare/v0.2.0...v0.2.1) - 2023-01-19